+++ projects/abraxas/voting-lib/src/lib/components/dialog/confirm-dialog/confirm-dialog.component.ts
@@ -5,7 +5,7 @@
  */
 
 import { Component, inject } from '@angular/core';
-import { MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
+import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog';
 
 @Component({
   templateUrl: './confirm-dialog.component.html',